Move common analytics functions out of analytics command
